Module2/worker/requirements.txt
mysql-connector-python
pika
psycopg2-binary